Current version of Psi+ is 1.5.1485
It is based on:
* psi: 8cdcca2c
* plugins: b03a4d0
* psimedia: a323296
* resources: a00b363

+++ iris/src/irisnet/noncore/cutestuff/httpconnect.h
@@ -31,7 +31,7 @@ public:
     ~HttpConnect();
 
     void setAuth(const QString &user, const QString &pass = "");
-    void connectToHost(const QString &proxyHost, int proxyPort, const QString &host, int port);
+    void connectToHost(const QString &proxyHost, quint16 proxyPort, const QString &host, int port);
 
     // from ByteStream
     void   close();
