+++ iris/src/irisnet/noncore/icetransport.h
@@ -40,7 +40,7 @@ public:
     virtual void stop() = 0;
 
     virtual bool       hasPendingDatagrams(int path) const                                                = 0;
-    virtual QByteArray readDatagram(int path, QHostAddress *addr, int *port)                              = 0;
+    virtual QByteArray readDatagram(int path, QHostAddress *addr, quint16 *port)                          = 0;
     virtual void       writeDatagram(int path, const QByteArray &buf, const QHostAddress &addr, int port) = 0;
     virtual void       addChannelPeer(const QHostAddress &addr, int port)                                 = 0;
